move `createResult()` into `RenderContext`
