+++ packages/astro/test/units/test-utils.js
@@ -201,7 +201,7 @@ export function createBasicPipeline(options = {}) {
 		options.routeCache ?? new RouteCache(options.logger ?? defaultLogger, mode),
 		options.site
 	);
-	pipeline.componentMetadata = () => {};
+	pipeline.componentMetadata = () => new Map;
 	pipeline.headElements = () => ({ scripts: new Set(), styles: new Set(), links: new Set() });
 	pipeline.resolve = id => id;
 	return pipeline;
